files names change, module directory create
navbar file inside modules create. navbar function add, exported and imported by index.js.
navbar working properly
