+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./css/style.css">
+    <link rel="stylesheet" href="./css/index.css">
     <title>Document</title>
 </head>
 <body>
@@ -66,7 +66,7 @@
         <h2>Copyright</h2>
         <p>...</p>
     </footer>
-    <script src="main.js"></script>
+    <script src="index.js" type="module"></script>
 </body>
 </html>
